requirements.txt
clickhouse-driver
apache-airflow-providers-clickhouse